refactor: Remove the time crate, as it duplicates a clock counter.
We don't need to simulate Time.h from arduino because we have a local clock in
`hr_now, min_now, sec_now`, as well as the incrementer `last_time`.

We don't need `digital_clock_display` because of `ufmt::uwriteln!`.

Modifying `get_pc_time` to return the clock instead of setting the `time` clock
and then fetching that again saves some space and simplifies the call.

Not too happy about the `try_into` though.