refactor: Extract LED macros into methods on a struct.
This avoids the Dynamic type which adds memory and runtime to the array, it
also looks more like the original code.  I also get to see the function in the
disassembly now.

I tried to change AllLEDsOff into some port manipulation like the original but
got stuck on sharing the pins and ports; the compiler didn't like that.