Convert the pin activation to use tri-state mode.
The original code sets pins to floating inputs in the `all_off` function,
which is not what this code was doing until now.

This change converts the activation function to move the LED pins to hi/lo
output when activating a specific LED, then moving them back to the pin array
when done, as deactivated.

This isn't exactly nice, and we eagerly await
https://github.com/Rahix/avr-hal/pull/205 to land so this code doesn't use
`mode::Dynamic` overhead, or the unsafe pointer access.